% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverse_events.R
\name{adverse_events}
\alias{adverse_events}
\title{This function scrapes the openFDA API for food, dietary supplement, and cosmetic adverse event reports}
\usage{
adverse_events(
  api_key,
  consumer.gender = NULL,
  consumer.age = NULL,
  consumer.age_unit = NULL,
  date_created = NULL,
  date_started = NULL,
  limit = NULL,
  outcomes = NULL,
  products.name_brand = NULL,
  products.industry_name = NULL,
  products.industry_code = NULL,
  reactions = NULL,
  report_number = NULL,
  search_mode = NULL
)
}
\arguments{
\item{api_key}{Your free api key from openFDA website}

\item{consumer.gender}{A way the FDA classifies a date}

\item{consumer.age}{Description of product}

\item{consumer.age_unit}{Date for which recall was initiated}

\item{date_created}{The company recalling the product}

\item{date_started}{The date the FDA issued the enforcement report for the product recall}

\item{limit}{The number of rows to return for that query}

\item{outcomes}{This gives the user flexibility to search for exact matches of inputs or any combination of inputs}

\item{products.name_brand}{The status of the recall}

\item{products.industry_name}{The date the recall was terminated}

\item{products.industry_code}{The date the recall was terminated}

\item{reactions}{The date the recall was terminated}

\item{report_number}{The date the recall was terminated}

\item{search_mode}{The way to search for the data}
}
\value{
A data frame with the returned results of the users query to the API
}
\description{
This function scrapes the openFDA API for food, dietary supplement, and cosmetic adverse event reports
}
\examples{
\dontrun{
adverse_events(api_key = api_key, outcomes = "Hospitalization", date_created = "2020")
adverse_events(api_key = api_key, products.industry_name = "Food", date_started = "2021")
adverse_events(api_key = api_key, products.name_brand = "Lucky Charms")
}
}
