% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usda_fsis.R
\name{get_usda_fsis}
\alias{get_usda_fsis}
\title{This function scrapes the USDA-FSIS website for food product recall data}
\usage{
get_usda_fsis(x)
}
\arguments{
\item{x}{Need to update the input functionality of the function}
}
\description{
This function scrapes the USDA-FSIS website for food product recall data
}
