% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bar_chart.R
\name{plot_bar_chart}
\alias{plot_bar_chart}
\title{Plot a bar chart}
\usage{
plot_bar_chart(data, x_col, y_col)
}
\arguments{
\item{data}{The data frame to plot}

\item{x_col}{The name of the column to use for the x-axis}

\item{y_col}{The name of the column to use for the y-axis}
}
\value{
A bar chart
}
\description{
Plot a bar chart
}
