% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_date.R
\name{recall_date}
\alias{recall_date}
\title{This function scrapes the FDA website for food product recall data}
\usage{
recall_date(
  api_key,
  center_classification_date = NULL,
  limit = NULL,
  product_description = NULL,
  recall_initiation_date = NULL,
  recalling_firm = NULL,
  report_date = NULL,
  search_mode = NULL,
  status = NULL,
  termination_date = NULL
)
}
\arguments{
\item{api_key}{Your free api key from FDA website}

\item{center_classification_date}{A way the FDA classifies a date}

\item{limit}{The number of rows to return for that query}

\item{product_description}{Description of product}

\item{recall_initiation_date}{Date for which recall was initiated}

\item{recalling_firm}{The company recalling the product}

\item{report_date}{The date the FDA issued the enforcement report for the product recall}

\item{status}{The status of the recall}

\item{termination_date}{The date the recall was terminated}
}
\description{
This function scrapes the FDA website for food product recall data
}
