% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recall_location.R
\name{recall_location}
\alias{recall_location}
\title{This function scrapes the FDA website for food product recall data}
\usage{
recall_location(
  api_key,
  limit = NULL,
  city = NULL,
  country = NULL,
  distribution_pattern = NULL,
  recalling_firm = NULL,
  state = NULL,
  status = NULL
)
}
\arguments{
\item{api_key}{Your free api key from FDA website}

\item{limit}{The number of rows to return for that query}

\item{city}{City where food was produced}

\item{country}{The country where the food was produced}

\item{distribution_pattern}{Locations where food was distributed to}

\item{recalling_firm}{The company recalling the product}

\item{state}{The U.S. state in which the recalling firm is located}

\item{status}{The status of the recall}
}
\description{
This function scrapes the FDA website for food product recall data
}
