% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverse_events.R
\name{adverse_events}
\alias{adverse_events}
\title{This function scrapes the FDA website for food product recall data}
\usage{
adverse_events(
  api_key,
  consumer.gender = NULL,
  consumer.age = NULL,
  consumer.age_unit = NULL,
  date_created = NULL,
  date_started = NULL,
  limit = NULL,
  outcomes = NULL,
  products.name_brand = NULL,
  products.industry_name = NULL,
  products.industry_code = NULL,
  reactions = NULL,
  report_number = NULL
)
}
\arguments{
\item{api_key}{Your free api key from FDA website}

\item{limit}{The number of rows to return for that query}

\item{center_classification_date}{A way the FDA classifies a date}

\item{product_description}{Description of product}

\item{recall_initiation_date}{Date for which recall was initiated}

\item{recalling_firm}{The company recalling the product}

\item{report_date}{The date the FDA issued the enforcement report for the product recall}

\item{search_mode}{This gives the user flexibility to search for exact matches of inputs or any combination of inputs}

\item{status}{The status of the recall}

\item{termination_date}{The date the recall was terminated}
}
\description{
This function scrapes the FDA website for food product recall data
}
