% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cdc.R
\name{get_cdc}
\alias{get_cdc}
\title{This function scrapes the USDA-FSIS website for food product recall data}
\usage{
get_cdc(x)
}
\arguments{
\item{x}{Need to update the input functionality of the function}
}
\description{
This function scrapes the USDA-FSIS website for food product recall data
}
