% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fda.R
\name{get_fda}
\alias{get_fda}
\title{This function scrapes the FDA website for food product recall data}
\usage{
get_fda(x)
}
\arguments{
\item{x}{Need to update the input functionality of the function}
}
\description{
This function scrapes the FDA website for food product recall data
}
